% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Markers.R
\name{Create_Markers}
\alias{Create_Markers}
\title{Create markers on bulk data}
\usage{
Create_Markers(
  bulk_data,
  colData,
  class_col = NULL,
  lg2FC = 1,
  p.adjust = 0.05,
  show_lg2FC = TRUE
)
}
\arguments{
\item{bulk_data}{A matrix of non-negative integers with row features and sample columns.}

\item{colData}{A data.frame with at least a single column. Rows of colData correspond to columns of bulk_data.}

\item{class_col}{The variable name of subclasses, and must be contained in the colData columns.}

\item{lg2FC}{In the DESeq differential expression analysis results, the cutoff value of lg2FC. The default value is 1.}

\item{p.adjust}{In the DESeq differential expression analysis results, the cutoff value of adjust P. The default value is 0.05.}

\item{show_lg2FC}{Select whether to show log2 fold changes. The default value is TRUE.}
}
\value{
This function returns a \code{data.frame} with rows are samples and the columns contain the following attributes:
\item{genes}{Differential genes screened out based on lg2FC and p.adjust.}
\item{class}{The subclass of bulk data to which the marker gene belongs.}
}
\description{
Based on the differential genes of bulk data (differential expression analysis was performed by DESeq2),
the feature vector of each subclass is established. The bulk data to be analyzed should have at least two subclasses.
You could customize the levels of subclasses, otherwise the subclasses will be automatically converted into a factor variable.
}
\examples{
# Please refer to vignette.
}
