% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPET_BarPlot.R
\name{PIPET_BarPlot}
\alias{PIPET_BarPlot}
\title{Bar plot of single cell data}
\usage{
PIPET_BarPlot(SC, group, prediction, palette = "Set1", ylim)
}
\arguments{
\item{SC}{A Seurat object of single cell data.}

\item{group}{A character, name of one metadata columns to group cells by.}

\item{prediction}{A character, the variable name of prediction, and must be contained in the metadata columns.}

\item{palette}{If a string, will use that named palette. If a number, will index into the list of palettes of appropriate type. The default value is "Set1".}

\item{ylim}{An integer, set limits for the y axis.}
}
\value{
An output of graphic or a list of ggplot objects.
}
\description{
Plot a bar chart and make the height of the bar proportional to the number of cells in each group.
}
\examples{
# Please refer to vignette.
}
