% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPET_DotPlot.R
\name{PIPET_DotPlot}
\alias{PIPET_DotPlot}
\title{Dimensional reduction plot of single cell data}
\usage{
PIPET_DotPlot(
  SC,
  colors,
  color_alpha = 1,
  reduction = "umap",
  group,
  pt.size = 1.2,
  label = FALSE,
  label.size = 5,
  title
)
}
\arguments{
\item{SC}{A Seurat object of single cell data.}

\item{colors}{A set of colors to map data values to.}

\item{color_alpha}{An alpha level in \link{0,1}. The default value is 1.}

\item{reduction}{Which dimensionality reduction to use. The default value is "umap".}

\item{group}{A character, name of one metadata columns to group cells by.}

\item{pt.size}{Adjust point size for plotting. The default value is 1.2.}

\item{label}{Whether to label the clusters. The default value is FALSE.}

\item{label.size}{Sets size of labels. The default value is 5}

\item{title}{Add a title to the plot.}
}
\value{
An output of graphic or a list of ggplot objects.
}
\description{
Graphs the output of a dimensional reduction technique on a 2D scatter plot where each point is a cell and it's
positioned based on the cell embeddings determined by the reduction technique. Cells are colored by subpopulations.
}
\examples{
# Please refer to vignette.
}
