% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPET_KM.R
\name{PIPET_KM}
\alias{PIPET_KM}
\title{Drawing survival curves for different risk groups}
\usage{
PIPET_KM(
  data,
  genes,
  surv_cols,
  coxph,
  time = "days",
  xlim = 120,
  break.time = 30,
  legend.title,
  legend.labs = c("HRisk", "LRisk"),
  pval = TRUE,
  conf.int = FALSE,
  palette = "simpsons"
)
}
\arguments{
\item{data}{A data frame with clinical information of the data to be analyzed.}

\item{genes}{A set of genes contained in the "data."}

\item{surv_cols}{A vector containing variable names for survival status and survival time. And they must be contained in the "data" columns.}

\item{coxph}{An object of class coxph representing the fit (output by coxph function).}

\item{time}{A string value to indicate how to store the survival time (days or months). The default value is days.}

\item{xlim}{An integer, set limits for the x axis.}

\item{break.time}{The numeric value controlling time axis breaks. Default value is 30.}

\item{legend.title}{Set a legend title.}

\item{legend.labs}{Character vector specifying legend labels. Used to replace the names of the strata from the fit.}

\item{pval}{It is used to control the display of p-values for comparisons between survival curves. The default value is TRUE.}

\item{conf.int}{It is used to control the display of confidence intervals around the survival curves in a Kaplan-Meier plot. The default value is FALSE.}

\item{palette}{The argument palette can be used to specify the color to be used for each group. The default value is "simpsons".}
}
\value{
An output of graphic or a list of ggsurvplot object.
}
\description{
It is a function to plot survival curves. First, calculate the risk score of each sample based on the results of multi-factor cox
regression analysis. The sample will then be divided into high and low risk groups based on the median value of the risk score. And plot
a list of survfit objects as generated by the Surv function.
}
\examples{
# Please refer to vignette.
}
