% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPET_MultiViolin.R
\name{PIPET_MultiViolin}
\alias{PIPET_MultiViolin}
\title{Multi-violin plot of single cell data}
\usage{
PIPET_MultiViolin(
  SC,
  features,
  preprocess = TRUE,
  col = "prediction",
  palette = "Set1",
  xlab = "Gene Expression",
  ylab = ""
)
}
\arguments{
\item{SC}{A Seurat object of single cell data.}

\item{features}{A set of features to plot.}

\item{preprocess}{Select whether to perform normalization and scale of data. The default value is TRUE.}

\item{col}{A character, the variable name of prediction, and must be contained in the metadata columns.}

\item{palette}{If a string, will use that named palette. If a number, will index into the list of palettes of appropriate type. The default value is "Set1".}

\item{xlab}{Set the label name of the x axis.}

\item{ylab}{Set the label name of the y axis.}
}
\value{
An output of graphic or a list of ggplot objects.
}
\description{
Draws a stacked violin plot with multiple modules of single cell data.
}
\examples{
# Please refer to vignette.
}
